gold/requirements.txt
django>=5.1
djangorestframework>=3.15
psycopg2-binary>=2.9
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
python-decouple>=3.8
whitenoise>=6.7
requests>=2.32
social-auth-app-django>=5.4
gunicorn>=23.0
dj-database-url>=2.3
pillow>=10.4
boto3==1.37.23
botocore==1.37.23
jmespath==1.0.1
s3transfer==0.11.4